+++ tests/TestStackHandler.py
@@ -14,8 +14,6 @@ class TestStackHandler(unittest.TestCase):
         TestUtilities.AssertStacksExists(["nginx_test"], True)
         handler.Remove(['nginx_test'])
         TestUtilities.AssertStacksExists(["nginx_test"], False)
-        handler.Remove(["infrastructure"], ignoreInfrastructure=False)
-        TestUtilities.AssertInfrastructureExists(False)
 
     def test_DeployOnlyValidStacks(self):
         handler: StackHandler = TestUtilities.CreateStackHandler(stacksFolder="tests/invalidTestStacks")
@@ -29,8 +27,6 @@ class TestStackHandler(unittest.TestCase):
         TestUtilities.AssertStacksExists(["nginx_test_invalid_volume"], False)
         handler.Deploy(['nginx_test_invalid_port'], verifyStacksOnDeploy=True)
         TestUtilities.AssertStacksExists(["nginx_test_invalid_port"], False)
-        handler.Remove(["infrastructure"], ignoreInfrastructure=False)
-        TestUtilities.AssertInfrastructureExists(False)
 
     def test_InitPrune(self):
         handler: StackHandler = TestUtilities.CreateStackHandler()
@@ -40,7 +36,6 @@ class TestStackHandler(unittest.TestCase):
         TestUtilities.AssertStacksExists(['nginx_test'], True)
         handler.Prune()
         TestUtilities.AssertStacksExists(['nginx_test'], False)
-        TestUtilities.AssertInfrastructureExists(False)
         stacks = handler.List()
         self.assertGreater(len(stacks), 0)
         TestUtilities.AssertStacksExists(stacks, False)
@@ -49,22 +44,21 @@ class TestStackHandler(unittest.TestCase):
         handler: StackHandler = TestUtilities.CreateStackHandler(offline=True)
         stacks = handler.List()
         self.assertTrue('nginx_test' in stacks)
-        self.assertTrue('infrastructure' in stacks)
 
     def test_List_online(self):
         handler: StackHandler = TestUtilities.CreateStackHandler(offline=False)
         stacks = handler.List()
         self.assertTrue('nginx_test_online' in stacks)
-        self.assertTrue('infrastructure_online' in stacks)
 
     def test_Verify(self):
         handler: StackHandler = TestUtilities.CreateStackHandler()
         self.assertFalse(handler.Verify())
-        self.assertTrue(handler.Verify(['infrastructure', 'ngint_test_digest']))
+        self.assertTrue(handler.Verify(['ngint_test_digest']))
 
     def test_Run(self):
         DockerComposeTools.DockerComposeBuild(["tests/testBatchStacks/docker-compose.batch.yml"])
-        handler: StackHandler = TestUtilities.CreateStackHandler(stacksFolder="tests/testBatchStacks")
+        handler: StackHandler = TestUtilities.CreateStackHandler(stacksFolder="tests/testBatchStacks",
+                                                                 swmInfrastructureFiles=["tests/testBatchStacks/swarm.management,yml"])
         handler.Init()
         os.environ['SHOULD_FAIL'] = 'false'
         os.environ['SHOULD_FAIL_2'] = 'false'